separate DOM from logic in todos and fixed ESlint errors
