+++ src/index.js
@@ -1,11 +1,12 @@
 import createEvent from './app/events';
 import { projectShow, projectForm } from './app/projectDom';
-import { todoForm } from './app/todos';
-import { createOptions } from './app/todoDom';
+import { todoForm, createOptions } from './app/todosDom';
+import { getProjectValues } from './app/projects';
+
 import './main.scss';
 
 createEvent();
-projectForm();
-projectShow();
+projectForm(getProjectValues());
+projectShow(getProjectValues());
 todoForm();
-createOptions();
+createOptions(getProjectValues());
